fix(review): resolve requested changes and issues from PR #36
