+++ .env.example
@@ -1,3 +1,2 @@
-BOOK_CHAPTER=book_name.rtl.md
 CHAPTER_NO=book_chapter_number
 GOOGLE_GENAI_API_KEY=your_gemini_api_key
