--- .vscode/settings.json
@@ -1,17 +1,0 @@
-{
-    "python.defaultInterpreterPath": "",
-    "emeraldwalk.runonsave": {
-        "commands": [
-            {
-                "match": "\\.md|.yml|.css$",
-                "cmd": "make build_book_via_docker"
-            }
-        ]
-    },
-    "[markdown]": {
-        "editor.fontFamily": "Noto Nastaliq Urdu",
-        "editor.fontSize": 42,
-        "editor.lineHeight": 2
-    },
-    "git.ignoreLimitWarning": true
-}
