+++ docs/audio-book.md
@@ -1,12 +1,50 @@
-# 🎧 Audiobook & Media Generation Guide
+# Text to emotional audio generation
 
-This guide explains how to use our standalone Node.js tool to convert raw text documents into high-quality audio files with realistic, emotional voice generation.
+## 🌟 Highlights
 
-## 🛠️ Step-by-Step Local Setup
+* **Purpose:** Convert a simple text into emtional audio.
+* **Voice support:** After choosing the speaker name from gemini supported speakers, change into audio-parsing api metadata.
+* **Application type:** It is console application, built to run directly in terminal.
+* **Format Agnostic:** Export directly to .wav .
 
-Follow these exact steps to run the audiobook generator locally:
+## ℹ️ Overview
 
-### 1. Install Dependencies
-Open your terminal at the root directory (`islamopedia-books/`) and install all required modules:
-```bash
-npm i
+This text to emotional audio generator is light-weight console application, which is designed to implement the Text-to-Speech (TTS) integration. Whether you are building an audio book, or want to make a speech with emotional multi-speaker voices, this tool is effective for you. This provide a clean and intuitive API to turn your string into high quality audio files.
+
+I built this because the solutions were either too bloated for simple use cases and required complex cloud dependencies. This tool aims to keep you audio generation local, private and lightning-fast.
+
+### ✍️ Developer
+
+Danish Ajmal - [Github Profile](https://github.com/DanishAjma1)
+
+## 🚀 Usage
+
+*Setup the envoironment variables in your .env file.*
+
+``` bash
+CHAPTER_NO=book_chapter_number
+GOOGLE_GENAI_API_KEY=your_gemini_api_key
+```
+
+
+*Open the terminal, navigate to **audio_generator** directory, install the dependencies and run the app.*
+
+``` bash
+cd scripts/audio-generator
+npm install
+npm start
+```
+
+## ⬇️ Installation
+
+Just clone the application!
+
+``` bash
+git clone https://github.com/DanishAjma1/islamopedia-books.git
+```
+
+And be sure your system must have node installed above 20x version.
+
+## 💭 Feedback and Contributing
+
+I have attached a discussion link, where I openly invite you to discuss about this repo's any part which could be enhance or bug you are tackling with. Tap Here - [Let's discuss](https://github.com/DanishAjma1/islamopedia-books/discussions/3)
