--- inputFiles/chap_1.rtl.md
@@ -1,8 +1,0 @@
-# ابتدائیہ اور تعارف
-
-<span class="quranic-text">بِسْمِ اللهِ الرَّحْمٰنِ الرَّحِيْمِ</span>
-
-## عرضِ ناشر
-
-<span class="arabic-text">الحمد لله الذی ھدٰنا لھذا وما کنّا لنھتدی لولآ ان ھدنا الله اللّٰھم صلّ علی محمّد ن النّبیّ الامّیّ وازواجه امھات المؤمنین وذریته واھل بیته کما صلیت علٰی ابرٰهیم انّك حمید مجید۔</span>
-
