+++ scripts/audio_generator/package.json
@@ -1,0 +1,41 @@
+﻿{
+  "name": "islamopedia-books",
+  "version": "1.0.0",
+  "description": "This is a helper repository of [Islamopedia](https://github.com/segullshairbutt/islamopedia/) to keep the backup of the books in text format. This will allow us to keep the books in a version control system and allow us to collaborate on the books easily. Currently finalized content can be seen live at: https://islamopedia-books.web.app",
+  "main": "index.ts",
+  "directories": {
+    "doc": "docs"
+  },
+  "scripts": {
+    "start": "node index.ts",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/DanishAjma1/islamopedia-books.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "type": "module",
+  "dependencies": {
+    "@google/genai": "^2.10.0",
+    "dotenv": "^17.4.2",
+    "ffmpeg-static": "^5.3.0",
+    "fluent-ffmpeg": "^2.1.3",
+    "fs": "^0.0.1-security",
+    "openai": "^6.42.0",
+    "pdf-parse": "^2.4.5",
+    "wav": "^1.0.2"
+  },
+  "devDependencies": {
+    "@jest/globals": "^30.4.1",
+    "@types/fluent-ffmpeg": "^2.1.28",
+    "@types/wav": "^1.0.4",
+    "jest": "^30.4.2"
+  },
+  "bugs": {
+    "url": "https://github.com/DanishAjma1/islamopedia-books/issues"
+  },
+  "homepage": "https://github.com/DanishAjma1/islamopedia-books#readme"
+}
