+++ scripts/audio_generator/segments/concatenation.ts
@@ -1,13 +1,14 @@
 import Ffmpeg from "fluent-ffmpeg";
 import { writeFileSync } from "fs";
 import path from "path";
+import { configDotenv } from "dotenv";
 
-export async function mergeFiles(
-  allAudioFiles: string[],
-  outputChapter: string,
-) {
+configDotenv();
+
+export async function mergeFiles(allAudioFiles: string[], outputChapter: string) {
   await new Promise<void>((resolve, reject) => {
-    const txtFilePath = "../outputAudios/tmp/filelist.txt";
+    const temDir = `../../../outputAudios/tmp`;
+    const txtFilePath = path.join(import.meta.dirname, temDir, `file_list.txt`);
 
     // 1. Create the text file
     const fileListContent = allAudioFiles
@@ -15,18 +16,18 @@ export async function mergeFiles(
       .join("\n");
     writeFileSync(txtFilePath, fileListContent);
 
-    console.log("Merging audio files into:", outputChapter);
+    console.log("\nMerging audio files into:", outputChapter);
     // 2. Run FFmpeg using the concat demuxer
     Ffmpeg()
       .input(txtFilePath)
       .inputOptions(["-f concat", "-safe 0"])
       .outputOptions("-c copy")
       .on("error", (err) => {
-        console.error("ffmpeg concat error:", err);
+        console.error("\nffmpeg concat error:", err);
         reject(err);
       })
       .on("end", () => {
-        console.log(`Saved chapter audio safely to ${outputChapter}`);
+        console.log(`\nSaved chapter audio safely to ${outputChapter}`);
         resolve();
       })
       .save(outputChapter);
