+++ scripts/audio_generator/segments/parseTextWithEmotions.ts
