+++ scripts/audio_generator/segments/textToSpeech.ts
